# openapi.json
{"openapi":"3.1.0","info":{"title":"Generate Backend Software Engineering Challenge 2025","version":"1.0.0"},"paths":{"/":{"get":{"summary":"API documentation.","responses":{"200":{"description":"API Documentation Page.","content":{"text/html":{"schema":{"type":"string"}}}},"500":{"description":"API Docs could not be found.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}},"/healthcheck":{"get":{"responses":{"200":{"description":"Server is Healthy!","content":{"application/json":{"schema":{"properties":{"message":{"enum":["OK"],"type":"string"}},"type":"object"}}}},"500":{"description":"Server is not Healthy!","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}},"/api/v1/member/register":{"summary":"Register your Northeastern email address and grab your token","post":{"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"description":"Must be a valid Northeastern email address.","format":"email","type":"string"},"nuid":{"description":"Valid nuid","format":"nuid","type":"string"}},"required":["email","nuid"],"type":"object"}}}},"responses":{"201":{"description":"Successfully registered your Northeastern email","content":{"application/json":{"schema":{"description":"Unique identifier associated with the registered northeastern email.","properties":{"id":{"format":"uuid","type":"string"}},"required":["id"],"type":"object"}}}},"400":{"description":"Invalid northeastern email address or nuid provided.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}},"409":{"description":"User has already been registered.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}},"500":{"description":"Internal server error.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}},"/api/v1/member":{"summary":"Get the associated id from the northeastern email address and nuid incase you forgot.","get":{"parameters":[{"name":"email","in":"query","description":"Northeastern email address.","required":true,"schema":{"format":"email","type":"string"}},{"name":"nuid","in":"query","description":"Northeastern NUID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successfully retrieved id from email and nuid","content":{"application/json":{"schema":{"description":"Unique identifier associated with the registered northeastern email.","properties":{"id":{"format":"uuid","type":"string"}},"required":["id"],"type":"object"}}}},"400":{"description":"Invalid northeastern email address or nuid provided.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}},"500":{"description":"Internal server error.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}},"/api/v1/challenge/{id}/aliens":{"get":{"parameters":[{"name":"id","in":"path","required":true,"schema":{"description":"Unique identifer for the registered participant.","example":"17aa5a93-73fc-4f8c-9977-2994481213be","format":"uuid","type":"string"}}],"responses":{"200":{"description":"Successfully gotten alien invasion data","content":{"application/json":{"schema":{"properties":{"waves":{"example":[{"type":"regular","count":50},{"type":"swift","count":10},{"type":"boss","count":2}],"items":{"properties":{"aliens":{"items":{"properties":{"type":{"enum":["regular","swift","boss"],"type":"string"},"count":{"minimum":0,"type":"integer"}},"required":["type","count"],"type":"object"},"type":"array"}},"type":"object"},"type":"array"},"alienTypes":{"properties":{"regular":{"properties":{"hp":{"enum":[2],"type":"integer"},"id":{"enum":[1],"type":"integer"},"atk":{"enum":[3],"type":"integer"}},"type":"object"},"swift":{"properties":{"hp":{"enum":[1],"type":"integer"},"id":{"enum":[2],"type":"integer"},"atk":{"enum":[5],"type":"integer"}},"type":"object"},"boss":{"properties":{"hp":{"enum":[10],"type":"integer"},"id":{"enum":[3],"type":"integer"},"atk":{"enum":[10],"type":"integer"}},"type":"object"}},"type":"object"},"budget":{"enum":[100],"type":"integer"},"wallDurability":{"enum":[100],"type":"integer"}},"type":"object"}}}},"401":{"description":"Invalid ID. Are you sure you are using the id that you got upon registration?","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}},"500":{"description":"Internal Server Error.","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}},"/api/v1/challenge/{id}/submit":{"post":{"parameters":[{"name":"id","in":"path","required":true,"schema":{"description":"Unique identifer for the registered participant.","example":"17aa5a93-73fc-4f8c-9977-2994481213be","format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"gunsPurchased":{"items":{"properties":{"type":{"enum":["turret","machineGun","rayGun"],"type":"string"}},"type":"object"},"type":"array"},"totalCost":{"type":"integer"},"assignments":{"items":{"properties":{"wave":{"minimum":1,"type":"integer"},"gunQueues":{"items":{"items":{"type":"integer"},"type":"array"},"type":"array"},"wallDurabilityRemaining":{"type":"integer"}},"type":"object"},"type":"array"}},"type":"object"}}}},"responses":{"200":{"description":"Verify submission against testing server oracle.","content":{"application/json":{"schema":{"oneOf":[{"properties":{"valid":{"type":"boolean"},"score":{"type":"integer"}},"type":"object"},{"properties":{"valid":{"type":"boolean"},"reason":{"type":"string"}},"type":"object"}]}}}},"400":{"description":"Malformed Submission","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"properties":{"message":{"type":"string"}},"required":["message"],"type":"object"}}}}}}}},"components":{"schemas":{}}}